[APR-248] chore: move DD metrics endpoint code into common (#382)
